backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-dotenv
httpx
loguru
dashscope
psycopg2-binary
